# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Operix API - The Operating System for Roblox Groups
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups for authenticated user
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group by ID
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group settings
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupBody"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Staff ─────────────────────────────────────────────────────────────
...
  /groups/{groupId}/dashboard:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get aggregated dashboard stats
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /groups/{groupId}/activity:
    get:
      operationId: getActivityFeed
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /groups/{groupId}/alerts:
    get:
      operationId: listAlerts
      tags: [dashboard]
      summary: List live alerts for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /groups/{groupId}/alerts/{alertId}/read:
    post:
      operationId: markAlertRead
      tags: [dashboard]
      summary: Mark an alert as read
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
components:
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        name:
          type: string
        robloxGroupId:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        staffCount:
          type: integer
        activeSessions:
          type: integer
        openCases:
          type: integer
        createdAt:
          type: string
      required: [id, clerkUserId, name, createdAt]
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
        robloxGroupId:
          type: string
        description:
          type: string
        logoUrl:
          type: string
      required: [name]
    UpdateGroupBody:
      type: object
      properties:
        name:
          type: string
        robloxGroupId:
          type: string
        description:
          type: string
        logoUrl:
          type: string

    # ── Staff ────────────────────────────────────────────────────────────
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        robloxUsername:
          type: string
        robloxId:
          type: string
        rank:
          type: string
        role:
          type: string
          enum: [admin, hr, moderator, staff]
        status:
          type: string
          enum: [active, inactive, flagged, suspended]
        totalActivityMinutes:
          type: integer
        sessionsCount:
          type: integer
        warningsCount:
          type: integer
        casesCount:
          type: integer
        lastSeenAt:
          type: string
        joinedAt:
          type: string
        avatarUrl:
          type: string
      required: [id, groupId, robloxUsername, role, status, totalActivityMinutes, sessionsCount, warningsCount, casesCount, joinedAt]
...
    AiInsight:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        staffId:
          type: integer
        staffUsername:
          type: string
        type:
          type: string
          enum: [inactivity, high_performance, anomaly, promotion_suggest, demotion_suggest, case_auto]
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        resolved:
          type: boolean
        createdAt:
          type: string
      required: [id, groupId, type, message, severity, resolved, createdAt]
...
    DashboardStats:
      type: object
      properties:
        totalStaff:
          type: integer
        activeStaff:
          type: integer
        inactiveStaff:
          type: integer
        flaggedStaff:
          type: integer
        activeSessions:
          type: integer
        totalSessionsToday:
          type: integer
        totalActivityMinutesToday:
          type: integer
        openCases:
          type: integer
        closedCasesThisWeek:
          type: integer
        pendingAlerts:
          type: integer
        automationRulesFired:
          type: integer
        topPerformers:
          type: array
          items:
            $ref: "#/components/schemas/StaffMember"
        recentInsights:
          type: array
          items:
            $ref: "#/components/schemas/AiInsight"
      required: [totalStaff, activeStaff, inactiveStaff, flaggedStaff, activeSessions, totalSessionsToday, totalActivityMinutesToday, openCases, closedCasesThisWeek, pendingAlerts, automationRulesFired, topPerformers, recentInsights]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        type:
          type: string
          enum: [session_start, session_end, case_created, case_closed, staff_flagged, alert_fired, staff_added]
        message:
          type: string
        staffUsername:
          type: string
        referenceId:
          type: integer
        createdAt:
          type: string
      required: [id, groupId, type, message, createdAt]
    Alert:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        type:
          type: string
          enum: [inactivity, abuse_detected, case_update, promotion_suggestion, session_anomaly]
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        staffId:
          type: integer
        staffUsername:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required: [id, groupId, type, message, severity, read, createdAt]