# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Operix API - The Operating System for Roblox Groups
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: number;
  groupId: number;
  type: ActivityItemType;
  message: string;
  staffUsername?: string;
  referenceId?: number;
  createdAt: string;
}