# lib/api-zod/src/generated/types/createGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Operix API - The Operating System for Roblox Groups
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGroupBody {
  name: string;
  robloxGroupId?: string;
  description?: string;
  logoUrl?: string;
}